feat/cleanup: Organized codebase and created a prusai3mk3 class
