+++ server-javascript/src/serial_communication/marlin_firmware/generic_serial_fabricator.md
